# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fiber Banking API
  version: 1.0.0
  description: |
    Dokumentasi OpenAPI untuk API perbankan sederhana berbasis Go Fiber.

    Semua pesan response menggunakan bahasa Indonesia dan format standar:
    - Sukses: `{ data: T, status: true, message: string }`
    - Error: `{ status: false, message: string }`
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register customer (customer)
      description: |
        Endpoint publik untuk mendaftarkan user baru dengan role otomatis `customer`.

        role = `customer`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              default:
                value:
                  name: Customer Baru
                  email: customerbaru@example.com
                  password: password123
      responses:
        "201":
          description: Registrasi berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      id: c3m2y8p9x0n4q7r1s6t5u8v2
                      name: Customer Baru
                      email: customerbaru@example.com
                      role: customer
                      account_number: "20260723153001u8v2"
                    status: true
                    message: Registrasi berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (admin, staff, customer)
      description: |
        Login user admin, staff, atau customer menggunakan email dan password.

        role = `admin`, `staff`, `customer`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              default:
                value:
                  email: customer@example.com
                  password: password123
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
                      user:
                        id: ccustomer00000000000000001
                        name: Customer Satu
                        email: customer@example.com
                        role: customer
                    status: true
                    message: Login berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Kirim OTP reset password (admin, staff, customer)
      description: |
        Mengirim OTP ke email via Resend. Jika `RESEND_API_KEY` kosong, OTP dicetak ke stdout untuk development.

        role = `admin`, `staff`, `customer`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
            examples:
              default:
                value:
                  email: customer@example.com
      responses:
        "200":
          description: OTP berhasil dikirim atau email tidak diungkap keberadaannya
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      otp_id: cotp000000000000000000001
                    status: true
                    message: OTP berhasil dikirim ke email
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verifikasi OTP reset password (admin, staff, customer)
      description: |
        Memverifikasi OTP reset password dan mengembalikan token reset password sementara.

        role = `admin`, `staff`, `customer`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
            examples:
              default:
                value:
                  email: customer@example.com
                  otp: "123456"
      responses:
        "200":
          description: OTP valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyOtpSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      email: customer@example.com
                      reset_token: reset-token-example
                    status: true
                    message: OTP valid
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Ubah password dengan token reset (admin, staff, customer)
      description: |
        Mengubah password baru setelah OTP terverifikasi. Token reset dikirim dari response verify OTP.

        role = `admin`, `staff`, `customer`
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            examples:
              default:
                value:
                  email: customer@example.com
                  reset_token: reset-token-example
                  password: passwordbaru123
                  verify_password: passwordbaru123
      responses:
        "200":
          description: Password berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptySuccessResponse"
              examples:
                success:
                  value:
                    status: true
                    message: Password berhasil diperbarui
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Gunakan header `Authorization: Bearer <access_token>`."
...
  responses:
    ErrorResponse:
      description: Contoh response error standar. Status code aktual dapat berupa 400, 401, 403, 404, 409, 502, atau 500 sesuai kasus.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            error:
              value:
                status: false
                message: Request tidak valid
...
  schemas:
    Role:
      type: string
      enum: [admin, staff, customer]
      example: customer
...
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: Request tidak valid
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Customer Baru
        email:
          type: string
          format: email
          example: customerbaru@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: password123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          format: password
          example: password123
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
    VerifyOtpRequest:
      type: object
      required: [email, otp]
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        otp:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
    ResetPasswordRequest:
      type: object
      required: [email, reset_token, password, verify_password]
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        reset_token:
          type: string
          example: reset-token-example
        password:
          type: string
          format: password
          minLength: 8
          example: passwordbaru123
        verify_password:
          type: string
          format: password
          minLength: 8
          example: passwordbaru123
    VerifyOtpData:
      type: object
      required: [email, reset_token]
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        reset_token:
          type: string
          example: reset-token-example
...
    UserData:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          example: ccustomer00000000000000001
        name:
          type: string
          example: Customer Satu
        email:
          type: string
          format: email
          example: customer@example.com
        role:
          $ref: "#/components/schemas/Role"
    RegisterData:
      allOf:
        - $ref: "#/components/schemas/UserData"
        - type: object
          required: [account_number]
          properties:
            account_number:
              type: string
              example: "1000000001"
    LoginData:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        user:
          $ref: "#/components/schemas/UserData"
    ForgotPasswordData:
      type: object
      nullable: true
      properties:
        otp_id:
          type: string
          example: cotp000000000000000000001
...
    EmptySuccessResponse:
      type: object
      required: [status, message]
      properties:
        data:
          nullable: true
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Berhasil
    RegisterSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/RegisterData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Registrasi berhasil
    LoginSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/LoginData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Login berhasil
    ForgotPasswordSuccessResponse:
      type: object
      required: [status, message]
      properties:
        data:
          $ref: "#/components/schemas/ForgotPasswordData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: OTP berhasil dikirim ke email
    VerifyOtpSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/VerifyOtpData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: OTP valid
...